requirements.txt
discord.py==2.6.0
flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dateutil
pytz